wheel1/requirements.txt
# Wheel系统 - 依赖包清单

# ========== 核心框架 ==========
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ========== LLM和嵌入服务 ==========
openai>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0

# ========== 文档处理 ==========
PyPDF2>=3.0.1
python-docx>=0.8.11
Pillow>=10.0.0
pytesseract>=0.3.10
paddleocr>=2.7.0.3
easyocr>=1.7.0

# ========== 向量数据库 ==========
pymilvus>=2.3.0
qdrant-client>=2.4.0
pinecone-client>=2.2.4
weaviate-client>=3.21.0

# ========== 传统数据库 ==========
psycopg2-binary>=2.9.9
pymongo>=4.5.0
sqlalchemy>=2.0.0

# ========== 缓存 ==========
redis>=5.0.1

# ========== 监控和日志 ==========
python-json-logger>=2.0.7
prometheus-client>=0.19.0

# ========== 工具和工具类 ==========
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# ========== 配置管理 ==========
python-dotenv>=1.0.0
pyyaml>=6.0

# ========== 测试 ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ========== 开发工具 ==========
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0

# ========== 可选：高级RAG和Agent ==========
# llama-index>=0.9.0
# autogen>=0.2.0
# dspy-ai>=2.0.0
